requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
baldertest~=0.1.0b0
setuptools~=57.0.0
mechanize~=0.4.7
requests~=2.27.1
